refactor: organize directory
