+++ packages/vite-node-miniflare/src/client/polyfills/node-vm.ts
