+++ packages/vite-node-miniflare/src/client/polyfills/usage-checker.ts
