+++ packages/vite-node-miniflare/src/client/worker-entry-script.ts
@@ -1,0 +1,3 @@
+// injected via esbuild define
+declare let __DEFINE_WORKER_ENTRY_SCRIPT: string;
+export const WORKER_ENTRY_SCRIPT = __DEFINE_WORKER_ENTRY_SCRIPT;
