--- packages/vite-node-miniflare/src/worker-script.ts
@@ -1,3 +1,0 @@
-// injected via esbuild define
-declare let __DEFINE_WORKER_SCRIPT: string;
-export const WORKER_SCRIPT = __DEFINE_WORKER_SCRIPT;
